Task 8.3 - solution with classes
