--- Lesson10HW/Lesson10HW/Source/Screen/SignUpScreen/SignUpViewController.swift
@@ -1,22 +1,0 @@
-//
-//  SignUpViewController.swift
-//  Lesson10HW
-//
-
-//
-
-import UIKit
-
-class SignUpViewController: UIViewController {
-    
-    @IBOutlet weak var contentView: SignUpView!
-    
-    @IBOutlet weak var scrollView: UIScrollView!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        subscrubeNotifications()
-        contentView.keyboardsTypes()
-    }
-}
-
